Trainings Service

+++ src/main/java/tc3005b224/amazonconnectinsights/repository/TrainingsRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import tc3005b224.amazonconnectinsights.models_sql.Trainings;
 
 @Repository
-public interface TrainingsRepository extends CrudRepository<Trainings, Integer>{
+public interface TrainingsRepository extends CrudRepository<Trainings, Long>{
 
 }
